Move dataset generator so that it does not depend on pytest
